+++ CHANGELOG.md
@@ -7,6 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.1] - 2026-02-18
+
+### Changed
+
+- Version bump to 0.1.1
+- Updated the panel, with File and Settings in the Menu Bar
+- Option ot save and load "current_view_settings"
+- Faster selection of variables with the option to select and de-select all heights or variables
+- Option to save only the variables thatn have been selected (although it requires some polishing to work properly)
+
+
+
+## [0.1.0] - 2026-02-01
+
 ### Added
 
 - v0.1.0 Initial project structure